requirements.txt
# Core libraries
pandas==2.0.3                  # For data handling and preprocessing
numpy==1.26.0                  # For numerical computations
sqlalchemy==2.0.20             # For database interactions
psycopg2-binary==2.9.6         # PostgreSQL driver (if using PostgreSQL)

# Machine Learning
scikit-learn==1.3.1            # For churn prediction and segmentation
joblib==1.3.2                  # For model saving/loading (optional)

